move detail screen to separate folder
