+++ example/lib/src/ui/device_detail/device_detail_screen.dart
